Separate app actions and clean up
