+++ frontend/src/business/overmind/App/index.ts
@@ -1,0 +1,3 @@
+import * as actions from "./actions";
+
+export { actions };
